Move some functions from sys::fs to api::fs
